requirements.txt
appnope==0.1.2
APScheduler==3.6.3
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.4
decorator==4.4.2
idna==2.10
ipykernel==5.3.4
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.7
jupyter-core==4.7.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pip==21.0.1
prompt-toolkit==3.0.8
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
python-dateutil==2.8.1
python-telegram-bot==13.2
pytz==2021.1
pyzmq==20.0.0
requests==2.25.1
setuptools==52.0.0
six==1.15.0
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
wcwidth==0.2.5
wheel==0.36.2
